Fix warning
git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@7947 3db68df2-c116-0410-a063-a993310a9797

+++ source/graphics/GameView.cpp
@@ -739,7 +739,7 @@ void CGameView::Update(float DeltaTime)
 	}
 	*/
 
-	m->RotateY.Wrap(-M_PI, M_PI);
+	m->RotateY.Wrap(-(float)M_PI, (float)M_PI);
 
 	// Update the camera matrix
 	SetupCameraMatrix(m, &m->ViewCamera.m_Orientation);
